requirements.txt
prettytable==0.7.2
PyExecJS==1.5.1
